// src/types/File/Comble/CombleList.ts
import ItemList from '@/types/File/ItemList';

interface CombleList {
    localTypeList: ItemList[];
    chauffageTypeList: ItemList[];
    batimentNatureList: ItemList[];
    chantierTypeList: ItemList[];
    niveauHabitationList: ItemList[];
    partieAIsolerList: ItemList[];
    puissanceCompteurList: ItemList[];
    accesCombleList: ItemList[];
    couvertureTypeList: ItemList[];
    charpenteTypeList: ItemList[];
    etatToitureList: ItemList[];
    isolationExistanteTypeList: ItemList[];
    rehausseTrappeTypeList: ItemList[];
    nbrAccesCombleList: ItemList[];
    nbrCompartimentsList: ItemList[];
}

export default CombleList;